requirements.txt
aniso8601==8.0.0
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
Flask-Login==0.5.0
flask-marshmallow==0.13.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
python-dotenv==0.13.0
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.17
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
